added basic layout
There are new files - layout and controller both contain classes.
Layout class, responsible for the ui, connects to the textual library and display the data it gets as a nice looking table. there is a basic tcss file, in the future will add more design there.
Controller class is the manager of the program, connects the logic layer (task manager) with the presentation layer (layout).